store a path prefix in the worktree; will be used for subtree checkout

+++ regress/worktree/worktree_test.c
@@ -83,6 +83,7 @@ remove_workdir(const char *worktree_path)
 	remove_meta_file(worktree_path, GOT_REF_HEAD);
 	remove_meta_file(worktree_path, GOT_WORKTREE_FILE_INDEX);
 	remove_meta_file(worktree_path, GOT_WORKTREE_REPOSITORY);
+	remove_meta_file(worktree_path, GOT_WORKTREE_PATH_PREFIX);
 	remove_meta_file(worktree_path, GOT_WORKTREE_FORMAT);
 	remove_got_dir(worktree_path);
 	rmdir(worktree_path);
@@ -125,7 +126,7 @@ worktree_init(const char *repo_path)
 	if (mkdtemp(worktree_path) == NULL)
 		goto done;
 
-	err = got_worktree_init(worktree_path, head_ref, repo);
+	err = got_worktree_init(worktree_path, head_ref, "/", repo);
 	if (err != NULL)
 		goto done;
 
@@ -135,6 +136,8 @@ worktree_init(const char *repo_path)
 	if (!check_meta_file_exists(worktree_path, GOT_WORKTREE_FILE_INDEX))
 		goto done;
 	if (!check_meta_file_exists(worktree_path, GOT_WORKTREE_REPOSITORY))
+		goto done;
+	if (!check_meta_file_exists(worktree_path, GOT_WORKTREE_PATH_PREFIX))
 		goto done;
 	if (!check_meta_file_exists(worktree_path, GOT_WORKTREE_FORMAT))
 		goto done;
@@ -205,7 +208,7 @@ worktree_init_exists(const char *repo_path)
 
 	if (!obstruct_meta_file(&path, worktree_path, GOT_REF_HEAD))
 		goto done;
-	err = got_worktree_init(worktree_path, head_ref, repo);
+	err = got_worktree_init(worktree_path, head_ref, "/", repo);
 	if (err != NULL && err->code == GOT_ERR_ERRNO && errno == EEXIST)
 		ok++;
 	unlink(path);
@@ -213,7 +216,7 @@ worktree_init_exists(const char *repo_path)
 
 	if (!obstruct_meta_file(&path, worktree_path, GOT_WORKTREE_FILE_INDEX))
 		goto done;
-	err = got_worktree_init(worktree_path, head_ref, repo);
+	err = got_worktree_init(worktree_path, head_ref, "/", repo);
 	if (err != NULL && err->code == GOT_ERR_ERRNO && errno == EEXIST)
 		ok++;
 	unlink(path);
@@ -221,7 +224,15 @@ worktree_init_exists(const char *repo_path)
 
 	if (!obstruct_meta_file(&path, worktree_path, GOT_WORKTREE_REPOSITORY))
 		goto done;
-	err = got_worktree_init(worktree_path, head_ref, repo);
+	err = got_worktree_init(worktree_path, head_ref, "/", repo);
+	if (err != NULL && err->code == GOT_ERR_ERRNO && errno == EEXIST)
+		ok++;
+	unlink(path);
+	free(path);
+
+	if (!obstruct_meta_file(&path, worktree_path, GOT_WORKTREE_PATH_PREFIX))
+		goto done;
+	err = got_worktree_init(worktree_path, head_ref, "/", repo);
 	if (err != NULL && err->code == GOT_ERR_ERRNO && errno == EEXIST)
 		ok++;
 	unlink(path);
@@ -229,7 +240,7 @@ worktree_init_exists(const char *repo_path)
 
 	if (!obstruct_meta_file(&path, worktree_path, GOT_WORKTREE_FORMAT))
 		goto done;
-	err = got_worktree_init(worktree_path, head_ref, repo);
+	err = got_worktree_init(worktree_path, head_ref, "/", repo);
 	if (err != NULL && err->code == GOT_ERR_ERRNO && errno == EEXIST)
 		ok++;
 	unlink(path);
@@ -241,9 +252,9 @@ done:
 	if (repo)
 		got_repo_close(repo);
 	free(gotpath);
-	if (ok == 4)
+	if (ok == 5)
 		remove_workdir(worktree_path);
-	return (ok == 4);
+	return (ok == 5);
 }
 
 #define RUN_TEST(expr, name) \
